+++ lib/puppet/provider/db_rcu/db_rcu.rb
@@ -32,6 +32,7 @@ Puppet::Type.type(:db_rcu).provide(:db_rcu) do
     Puppet.debug 'rcu_status'
 
     oracle_home             = resource[:oracle_home]
+    sys_user                = resource[:sys_user]
     sys_password            = resource[:sys_password]
     user                    = resource[:os_user]
     prefix                  = resource[:name]
@@ -54,7 +55,7 @@ EOS
     FileUtils.chmod(0555, tmpFile.path)
 
     Puppet.debug "rcu for prefix #{prefix} execute SQL"
-    output = `su - #{user} -c 'export ORACLE_HOME=#{oracle_home};LD_LIBRARY_PATH=#{oracle_home}/lib;#{oracle_home}/bin/sqlplus \"sys/#{sys_password}@//#{db_server}/#{db_service} as sysdba\" @#{tmpFile.path}'`
+    output = `su - #{user} -c 'export ORACLE_HOME=#{oracle_home};LD_LIBRARY_PATH=#{oracle_home}/lib;#{oracle_home}/bin/sqlplus \"#{sys_user}/#{sys_password}@//#{db_server}/#{db_service} as sysdba\" @#{tmpFile.path}'`
     raise ArgumentError, "Error executing puppet code, #{output}" if $? != 0
 
     if FileTest.exists?("/tmp/check_rcu_#{prefix}2.txt")