+++ lib/puppet/type/db_rcu.rb
@@ -52,6 +52,12 @@ module Puppet
       EOT
     end
 
+    newparam(:sys_user) do
+      desc <<-EOT
+        The sys username for the RCU check/install.
+      EOT
+    end
+
     newparam(:sys_password) do
       desc <<-EOT
         The sys password for the RCU check/install.