+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.33
 - Small Suse fix for the autostart service
 - new installdb attribute cleanup_installfile
+- option to provide the sys username for RCU
 
 ## 1.0.32
 - be able to provide a listener name for starting the oracle listener ( manifest and custom type)